preupdate_hook: move convertAssign code to preupdate_hook file
Move the convertAssign logic from convert.go to
sqlite3_opt_preupdate_hook.go since that is the
only place it is used.
